+++ ovp_users/tests/test_views/profile.py
@@ -23,7 +23,6 @@ class ProfileTestCase(TestCase):
     response = create_user_with_profile(profile=self.profile)
     self.assertTrue(response.data['profile']['full_name'] == self.profile['full_name'])
     self.assertTrue(response.data['profile']['about'] == self.profile['about'])
-    self.assertTrue(response.data['profile']['public'] == True)
     self._assert_causes_and_skills_in_response(response)
     return response
 
@@ -106,7 +105,6 @@ class ProfileTestCase(TestCase):
     self.assertTrue(response.data['profile']['causes'][0]['id'] == ["Cause with 'id' 999 does not exist."])
     self.assertTrue(response.data['profile']['causes'][1]['id'] == ["Cause with 'id' 998 does not exist."])
 
-
   def _assert_causes_and_skills_in_response(self, response):
     self.assertTrue(response.data['profile']['skills'][0]['id'] == 1)
     self.assertTrue('name' in response.data['profile']['skills'][0])
@@ -126,14 +124,12 @@ class ProfileTestCase(TestCase):
         'about': 'New about',
         'causes': [{'id': 3}, {'id': 4}],
         'skills': [{'id': 3}, {'id': 4}],
-        'public': False,
       }
     }
 
     response = self.client.patch(reverse('user-current-user'), data, format="json")
     self.assertTrue(response.data['profile']['full_name'] == data['profile']['full_name'])
     self.assertTrue(response.data['profile']['about'] == data['profile']['about'])
-    self.assertTrue(response.data['profile']['public'] == False)
 
     self.assertTrue(response.data['profile']['skills'][0]['id'] == 3)
     self.assertTrue(response.data['profile']['skills'][1]['id'] == 4)